Change (most) experience expressions to use ML/3 rather than ML/4. Read substats in api into a long rather than an int.
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@14827 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/request/CharSheetRequest.java
@@ -489,13 +489,13 @@ public class CharSheetRequest
 		throws JSONException
 	{
 		int muscle = JSON.getInt( "muscle" );
-		int rawmuscle = JSON.getInt( "rawmuscle" );
+		long rawmuscle = JSON.getLong( "rawmuscle" );
 
 		int mysticality = JSON.getInt( "mysticality" );
-		int rawmysticality = JSON.getInt( "rawmysticality" );
+		long rawmysticality = JSON.getLong( "rawmysticality" );
 
 		int moxie = JSON.getInt( "moxie" );
-		int rawmoxie = JSON.getInt( "rawmoxie" );
+		long rawmoxie = JSON.getLong( "rawmoxie" );
 
 		KoLCharacter.setStatPoints( muscle, rawmuscle,
 					    mysticality, rawmysticality,
